packages/forge-data/pyproject.toml
[project]
name = "forge-data"
version = "0.1.0"
description = "Data curation pipeline: download, embed, filter, caption"
requires-python = ">=3.11,<3.13"
dependencies = [
    "forge-core",
    "webdataset>=0.2.86",
    "pillow>=11.0",
    "pyarrow>=18.0",
    "numpy>=1.26",
]

[project.optional-dependencies]
# uv add forge-data[download]
download = [
    "img2dataset>=1.45",
]
# uv add forge-data[caption]
caption = [
    "transformers>=4.47",
]
all = [
    "forge-data[download,caption]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/forge"]

[tool.uv.sources]
forge-core = { workspace = true }
